Add a _working_ early loader
