+++ .gitignore
@@ -125,9 +125,9 @@ freeline/
 freeline_project_description.json
 
 # CityQuestWebUI build files
-CityQuestWebUI/web/*.js
-CityQuestWebUI/web/*.js.map
-CityQuestWebUI/web/libs/*
+CityQuestWebUI/web/**/*.js
+CityQuestWebUI/web/**/*.js.map
+CityQuestWebUI/web/**/libs/*
 CityQuestWebUI/web/**/*.kjsm
 
 # CityQuestWebUI google maps API key
