+++ CityQuestWebUI/.idea/gradle.xml
@@ -9,6 +9,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/earlyloader" />
           </set>
         </option>
         <option name="useAutoImport" value="true" />
