+++ CityQuestWebUI/earlyloader/src/main/kotlin/ru.spbau.cityquest.webui.earlyloader/GoogleMapsAPIKey.kt
@@ -1,0 +1,3 @@
+package ru.spbau.cityquest.webui.earlyloader
+
+const val GOOGLE_MAPS_API_KEY : String = "AIzaSyCZ3dHdfUC8xiMlVArEx2ugypCd0f5lheY"
