+++ CityQuestWebUI/settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'CityQuestWebUI'
 
+include 'earlyloader'
