+++ CityQuestWebUI/src/main/kotlin/ru/spbau/cityquest/webui/questeditor/main.kt
@@ -5,23 +5,16 @@ import google.maps.LatLng
 import google.maps.MapOptions
 import kotlin.browser.document
 
-fun performChecks() {
-    if (document.head == null)
-        println("Warning: document.head is null, the QuestEditor may not work properly.")
-}
+//val optionsG = MapOptions(LatLng(59.9342802, 30.3350986), 12)
 
-@JsName("initMapCallback")
-fun initMapCallback() {
+fun createMap() {
     val node = document.getElementById("map")
     val options = MapOptions(LatLng(59.9342802, 30.3350986), 12)
+    println(JSON.stringify(options.toJson()))
     val map = KtGoogleMap(node, options)
 }
 
 fun main(args: Array<String>) {
     println("CityQuest Web UI in Kotlin is at your service!")
-
-    performChecks()
-
-    createGoogleMapsLoader()
-    println("Google Maps loader has been added to the HTML DOM.")
+    createMap()
 }
