+++ CityQuestWebUI/web/QuestEditor.html
@@ -3,10 +3,22 @@
 <head>
     <meta charset="UTF-8">
     <title>Quest Editor &mdash; CityQuest</title>
-    <script type="text/javascript" src="libs/kotlin.js"></script>
-    <script type="text/javascript" src="EarlyLoader.js"></script>
+
+    <style>
+        #map {
+            height: 100%;
+        }
+        html, body {
+            height: 100%;
+            margin: 0;
+            padding: 0;
+        }
+    </style>
+
+    <script type="text/javascript" src="EarlyLoader/libs/kotlin.js"></script>
+    <script type="text/javascript" src="EarlyLoader/earlyloader.js"></script>
 </head>
 <body>
-
+    <div id="map"></div>
 </body>
 </html>
